+++ parse.c
@@ -2095,8 +2095,8 @@ int do_squeeze_entry (name_list **list,	/* squeeze or dont-squeeze list */
 
 	if (!sinfo) {
 	    twmrc_error_prefix();
-	    fprintf (stderr, "unable to allocate %d bytes for squeeze info\n",
-		     sizeof(SqueezeInfo));
+	    fprintf (stderr, "unable to allocate %lu bytes for squeeze info\n",
+		     (unsigned long) sizeof(SqueezeInfo));
 	    return (1);
 	}
 	sinfo->justify = justify;
